+++ src/main.rs
@@ -91,7 +91,7 @@ fn parse_proc_stat(content: &str) -> Result<ProcStat<'_>, ProcStatError> {
 
     let open = content.find('(').ok_or(ProcStatError::InvalidFormat)?;
     let close = content[open + 1..]
-        .find(')')
+        .rfind(')')
         .map(|i| open + 1 + i)
         .ok_or(ProcStatError::InvalidFormat)?;
 
@@ -125,15 +125,15 @@ fn parse_proc_stat(content: &str) -> Result<ProcStat<'_>, ProcStatError> {
     Ok(ProcStat{pid, comm, state, ppid})
 }
 
-fn get_process_name(pid: i32) -> Result<String, String> {
+fn get_process_name(pid: &i32) -> Result<String, String> {
     let path = format!("/proc/{}/stat", pid);
     let contents = fs::read_to_string(&path)
    	        .map_err(|_| format!("Could not read {}", path))?;
     let proc_stat = parse_proc_stat(&contents).map_err(|e| {
         format!(
             "Unsupported /proc/{}/stat format ({:?}). \
-             Your system is currently not supported. \
-             Please open an issue with your kernel version.",
+             Either the process name is or your system is currently not supported. \
+             Please open an issue with the complete /proc/pid/stat content and your kernel version.",
             pid, e
         )
     })?;
@@ -204,13 +204,15 @@ fn read_rss_kb(pid: &i32) -> u64{
 
 fn find_descendants(
     parent_of: &HashMap<i32, i32>,
-    target_pid: i32,
+    target_pids: &Vec<i32>,
 ) -> HashSet<i32> {
 	// Given a mapping of pid -> ppid and a target pid,
-	// return all descendants of the target (including the target itself)
+	// return all descendants of the targets (including the targets themself)
     let mut descendants = HashSet::new();
-    descendants.insert(target_pid);
-	let mut found_new: bool;
+    for pid in target_pids{
+    	descendants.insert(*pid);
+    }
+    let mut found_new: bool;
     loop {
     	found_new = false;
         for (&pid, &ppid) in parent_of.iter() {
@@ -289,6 +291,19 @@ fn setup_output(spec: OutputSpec) -> io::Result<Output> {
 }
 
 
+fn get_pids_from_name(name: String) -> Vec<i32>{
+	let mut result_pids: Vec<i32> = Vec::new();
+	let all_pids = list_processes();
+	for pid in all_pids{
+		let x = get_process_name(&pid).unwrap();
+		if x == name{
+			result_pids.push(pid);
+		}  
+	}
+	result_pids
+}
+
+
 
 #[derive(Debug)]
 #[allow(dead_code)]
@@ -303,7 +318,7 @@ struct Args{
 	final_flag: bool,
 	hz: u64,
 	output: OutputSpec,
-	target_pid: i32,
+	target_pids: Vec<i32>,
 }
 
 
@@ -313,12 +328,23 @@ fn parse_args(args: &[String]) -> Result<Args, ParseArgError> {
     let mut hz = 1;
     let mut output = OutputSpec::Stdout;
     let mut pid = None;
+    let mut name = None;
+    let mut target_pids: Vec<i32> = Vec::new();
 
     let mut iter = args.iter().skip(1).peekable(); // skip program name
 
     while let Some(arg) = iter.next() {
         match arg.as_str() {
-            "--help" => help_flag = true,
+            "--help" => {
+            	help_flag = true;
+            	return Ok(Args {
+            	        help_flag,
+            	        final_flag,
+            	        hz,
+            	        output,
+            	        target_pids,
+            	    });
+            }
             "--final" => final_flag = true,
             "--hertz" => {
                 let value = iter.next().ok_or(ParseArgError::MissingValue("hertz"))?;
@@ -331,21 +357,29 @@ fn parse_args(args: &[String]) -> Result<Args, ParseArgError> {
                 let value = iter.next().ok_or(ParseArgError::MissingValue("output-file"))?;
                 output = OutputSpec::File(PathBuf::from(value));
             }
+            "--name" => {
+            	let value = iter.next().ok_or(ParseArgError::MissingValue("name"))?;
+            	name = Some("(".to_string() + value + ")");
+            }
             other => {
                 // assume PID if numeric
                 pid = Some(other.parse().map_err(|_| ParseArgError::InvalidValue("pid"))?);
             }
         }
     }
-
-    let target_pid = pid.ok_or(ParseArgError::MissingValue("pid"))?;
+    if let Some(name_val) = name {
+        target_pids.append(&mut get_pids_from_name(name_val));
+    } else {
+        let target_pid = pid.ok_or(ParseArgError::MissingValue("pid"))?; // accept only one pid from raw args
+        target_pids.push(target_pid);
+    }
 
     Ok(Args {
         help_flag,
         final_flag,
         hz,
         output,
-        target_pid,
+        target_pids,
     })
 }
 
@@ -367,6 +401,7 @@ fn main() {
 			Options:\n\
 			--hertz int, the desired number of iterations per second\n\
 			--output-file str, the file path where to write the output (stdout if absent)\n\
+			--name str, sum up the memory usage of all processes with this name (disable the <pid> argument)\n\
 			Flags:\n\
 			--final, display only 1 line with the max value",
     		version
@@ -376,7 +411,7 @@ fn main() {
     
 	let sleep_duration: u64 = 1000 / args.hz;
 
-    let process_name = match get_process_name(args.target_pid) {
+    let process_name = match get_process_name(args.target_pids.first().unwrap()) {
 	    Ok(name) => name,
 	    Err(msg) => {
 	        eprintln!("memimpact error: {}", msg);
@@ -395,7 +430,7 @@ fn main() {
 	if !args.final_flag{
 	    write_output(
 	    	&mut output,
-	    	format!("Tracking memory usage of PID {} {}\n", args.target_pid, process_name)
+	    	format!("Tracking memory usage of PID {} {}\n", args.target_pids.first().unwrap(), process_name)
 	    );
 	}
 
@@ -403,11 +438,18 @@ fn main() {
     let mut current: u64;
 
     loop {
+    	let mut stop_loop = false;
         let mapping = get_map_pid_to_ppid();
-        if !mapping.contains_key(&(args.target_pid)){
+        for pid in &args.target_pids{
+        	 if !mapping.contains_key(pid){
+        	 	stop_loop = true;
+	        	break;
+    	    } 
+        }
+        if stop_loop{
         	break;
         }
-        let target_descendants = find_descendants(&mapping, args.target_pid);
+        let target_descendants = find_descendants(&mapping, &args.target_pids);
         current = target_descendants.iter().map(read_rss_kb).sum();
         
         max = max.max(current);
@@ -418,7 +460,7 @@ fn main() {
 	        	&mut output,
 	        	format!(
 	        		"PID {} {}: current {}, max {}\n",
-	        		args.target_pid,
+	        		args.target_pids.first().unwrap(),
 	        		process_name,
 	        		display_current,
 	        		display_max
@@ -430,7 +472,7 @@ fn main() {
     let display_max = format_memory(max);
     write_output(
     	&mut output,
-    	format!("PID {} {}: max {}\n", args.target_pid, process_name, display_max )
+    	format!("PID {} {}: max {}\n", args.target_pids.first().unwrap(), process_name, display_max )
     );
 }
 
@@ -459,6 +501,16 @@ mod tests {
         assert_eq!(actual, expected);
     }
 
+    #[test]
+    fn test_parse_proc_stat_with_paranthesis_in_name() {
+    	// real world test case
+        let input = "3674 ((sd-pam)) S 3672 3672 3672 0 -1 4194624 49 0 0 0 0 0 0 0 20 0 1 0 4058 17170432 450 18446744073709551615 1 1 0 0 0 0 0 4096 0 0 0 0 17 8 0 0 0 0 0 0 0 0 0 0 0 0 0";
+        let actual = parse_proc_stat(input).unwrap();
+
+        let expected = ProcStat{pid: 3674, comm: "((sd-pam))", state: ProcessState::S, ppid: 3672};
+        assert_eq!(actual, expected);
+    }
+
 
     #[test]
     fn test_parse_proc_stat_invalid_missing_parens() {
@@ -476,7 +528,7 @@ mod tests {
         map.insert(4, 2);
         map.insert(5, 4);
 
-        let descendants = find_descendants(&map, 1);
+        let descendants = find_descendants(&map, &vec![1]);
 
         let expected: HashSet<i32> = [1, 2, 3, 4, 5].into_iter().collect();
         assert_eq!(descendants, expected);
@@ -488,7 +540,7 @@ mod tests {
         map.insert(2, 1);
         map.insert(3, 1);
 
-        let descendants = find_descendants(&map, 2);
+        let descendants = find_descendants(&map, &vec![2]);
 
         let expected: HashSet<i32> = [2].into_iter().collect();
         assert_eq!(descendants, expected);
@@ -550,7 +602,7 @@ mod tests {
         assert_eq!(parsed.final_flag, false);
         assert_eq!(parsed.hz, 1);
         matches!(parsed.output, OutputSpec::Stdout);
-        assert_eq!(parsed.target_pid, 1234);
+        assert_eq!(parsed.target_pids, vec![1234]);
     }
 
     #[test]
@@ -568,7 +620,7 @@ mod tests {
         assert!(parsed.final_flag);
         assert!(!parsed.help_flag);
         assert_eq!(parsed.hz, 10);
-        assert_eq!(parsed.target_pid, 4321);
+        assert_eq!(parsed.target_pids, vec![4321]);
 
         match parsed.output {
             OutputSpec::File(path) => assert_eq!(path, PathBuf::from("out.txt")),
@@ -578,11 +630,10 @@ mod tests {
 
     #[test]
     fn help_flag_only() {
-        let argv = args(&["memimpact", "--help", "999"]);
+        let argv = args(&["memimpact", "--help"]);
 
         let parsed = parse_args(&argv).unwrap();
         assert!(parsed.help_flag);
-        assert_eq!(parsed.target_pid, 999);
     }
 
     #[test]
@@ -683,7 +734,7 @@ mod tests {
 
         assert!(parsed.final_flag);
         assert_eq!(parsed.hz, 5);
-        assert_eq!(parsed.target_pid, 5678);
+        assert_eq!(parsed.target_pids, vec![5678]);
     }
 
     #[test]
@@ -699,6 +750,6 @@ mod tests {
 
         assert!(parsed.final_flag);
         assert_eq!(parsed.hz, 5);
-        assert_eq!(parsed.target_pid, 5678);
+        assert_eq!(parsed.target_pids, vec![5678]);
     }
 }